feat(bindings/java): support get enabled services
