+++ bindings/java/src/lib.rs
@@ -41,6 +41,7 @@ use tokio::runtime::Runtime;
 mod blocking_operator;
 mod error;
 mod operator;
+mod utility;
 
 pub(crate) type Result<T> = std::result::Result<T, error::Error>;
